v3
- `m.Keys` returns an `iter.Seq` iterator instead of a list of keys.
- Added `m.Values` which returns an `iter.Seq` iterator.
- Renamed Iterator method to `AllFromFront`.
- Renamed ReverseIterator to `AllFromBack`.
- Added `All` method aliased to `AllFromFront`.
- Updated minimum version to 1.23
- Moved methods into main file as they no longer need to be constrained by a build tag
- Updated tests
