+++ v3/go.mod
@@ -1,11 +1,11 @@
-module github.com/elliotchance/orderedmap/v2
+module github.com/elliotchance/orderedmap/v3
 
-go 1.18
+go 1.23.0
+
+require github.com/stretchr/testify v1.7.1
 
 require (
 	github.com/davecgh/go-spew v1.1.0 // indirect
 	github.com/pmezard/go-difflib v1.0.0 // indirect
-	github.com/stretchr/testify v1.7.1 // indirect
-	golang.org/x/exp v0.0.0-20220321173239-a90fa8a75705 // indirect
 	gopkg.in/yaml.v3 v3.0.0-20200313102051-9f266ea9e77c // indirect
 )
